onnx: move the execution plan out of loader/ into plan/
Loading and planning are separate concerns that shared a module. The
loader decodes a model; the plan decides how to run it — classifying each
node, selecting fusions, resolving convolution parameters, prepacking
weights. Keeping both in loader/ is why "the load-time graph optimizer"
was a doc comment on a function inside the file that parses protobuf.

Behaviour-preserving move. NodeAction, NodeKind, ConvParams,
RuntimeModelIndex and the two FusedPwDwPwReduce types go to plan/mod.rs;
runtime_index.rs becomes plan/build.rs. loader/mod.rs drops from 691
lines to 375 and is now about parsing a model, which is what its name
says. build.rs gains explicit imports rather than inheriting the loader's
via `use super::*`, so what it actually depends on is visible.

No logic changed and no function was split — this is the module boundary
Stage 3 needs before the 1771-line builder is broken into phases, done
separately so that diff is reviewable as logic rather than as movement.

Validation: 233 yscv-onnx tests, 2305 workspace, unchanged. cargo fmt,
clippy --workspace --all-targets --all-features -D warnings,
check-doc-counts.sh.
