+++ crates/yscv-onnx/src/runner/conv/fused.rs
@@ -29,8 +29,8 @@ pub(crate) fn exec_fused_dw_pw(
     env: &mut TensorEnv,
     dw_activation: yscv_kernels::Activation,
     pw_activation: yscv_kernels::Activation,
-    dw_params: Option<&crate::loader::ConvParams>,
-    pw_params: Option<&crate::loader::ConvParams>,
+    dw_params: Option<&crate::plan::ConvParams>,
+    pw_params: Option<&crate::plan::ConvParams>,
     dw_input_ids: &[Option<usize>],
     remaining_uses: &mut [usize],
     output_id_mask: &[bool],
@@ -502,8 +502,8 @@ pub(crate) fn exec_fused_pw_dw(
     env: &mut TensorEnv,
     pw_activation: yscv_kernels::Activation,
     dw_activation: yscv_kernels::Activation,
-    pw_params: Option<&crate::loader::ConvParams>,
-    dw_params: Option<&crate::loader::ConvParams>,
+    pw_params: Option<&crate::plan::ConvParams>,
+    dw_params: Option<&crate::plan::ConvParams>,
     pw_input_ids: &[Option<usize>],
     remaining_uses: &mut [usize],
     output_id_mask: &[bool],
@@ -1146,10 +1146,10 @@ pub(crate) fn exec_fused_pw_dw_pw_reduce(
     pw_expand_activation: yscv_kernels::Activation,
     dw_activation: yscv_kernels::Activation,
     pw_reduce_activation: yscv_kernels::Activation,
-    pw_expand_params: Option<&crate::loader::ConvParams>,
-    dw_params: Option<&crate::loader::ConvParams>,
+    pw_expand_params: Option<&crate::plan::ConvParams>,
+    dw_params: Option<&crate::plan::ConvParams>,
     pw_expand_input_ids: &[Option<usize>],
-    prepacked_pw_reduce: &crate::loader::FusedPwDwPwReduceWeights,
+    prepacked_pw_reduce: &crate::plan::FusedPwDwPwReduceWeights,
     residual_meta: Option<(&OnnxNode, u8, u8, Option<&OnnxNode>)>,
     dw_kernel_size: u8,
     remaining_uses: &mut [usize],
@@ -1577,8 +1577,8 @@ pub(crate) fn exec_fused_pw_dw_then_reduce_fallback(
     pw_expand_activation: yscv_kernels::Activation,
     dw_activation: yscv_kernels::Activation,
     pw_reduce_activation: yscv_kernels::Activation,
-    pw_expand_params: Option<&crate::loader::ConvParams>,
-    dw_params: Option<&crate::loader::ConvParams>,
+    pw_expand_params: Option<&crate::plan::ConvParams>,
+    dw_params: Option<&crate::plan::ConvParams>,
     pw_expand_input_ids: &[Option<usize>],
     residual_meta: Option<(&OnnxNode, u8, u8, Option<&OnnxNode>)>,
     remaining_uses: &mut [usize],
