+++ crates/yscv-onnx/src/runner/mod.rs
@@ -12,7 +12,8 @@ pub(crate) use yscv_tensor::{DType, Tensor};
 
 pub(crate) use crate::attr::Attr;
 pub(crate) use crate::error::OnnxError;
-pub(crate) use crate::loader::{NodeKind, OnnxAttribute, OnnxModel, OnnxNode};
+use crate::loader::{OnnxAttribute, OnnxModel, OnnxNode};
+pub(crate) use crate::plan::NodeKind;
 
 mod compare;
 mod conv;
